Handle default data properly
